Created framework for measuring performance with minimal overhead.
Alternative to using web inspector to measure performance.
Enabled under PERFORMANCE_TRACING flag.  If this flag isn't enabled this
code introduces no overhead.  Performance is measured with tags like this:

  PERFORMANCE_START(WTF::PerformanceTrace::ExecuteJavaScript, L"Script");
  JSValue returnValue = JSMainThreadExecState::evaluate(exec, jsSourceCode, shell, &evaluationException);
  InspectorInstrumentation::didEvaluateScript(cookie);
  PERFORMANCE_END(WTF::PerformanceTrace::ExecuteJavaScript);

Client can set output handler which is either invoked automatically when
the buffer is full or is invoked when the client calls WTF::PerformanceTrace::dumpEvents().